Improved the set of random data entered in the DB (seed operation)

+++ MatchBox.Contracts/Abstract/EntityBase.cs
@@ -6,6 +6,6 @@ namespace MatchBox.Contracts
 {
     public abstract class EntityBase
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
     }
 }
